+++ lib/arel/extensions.rb
@@ -30,6 +30,7 @@ require 'arel/extensions/distinct_from'
 require 'arel/extensions/not_distinct_from'
 require 'arel/extensions/null_if'
 require 'arel/extensions/similar'
+require 'arel/extensions/not_similar'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
